Move prove() to trace gen (#722)
